+++ CHANGELOG.md
@@ -1,2 +1,3 @@
+- Improve error message raised when `firebase init hosting:github` fails due to max number of keys limit for a service account. (#6145)
 - Fixed bug where `functions:secrets:\*` family of commands did not work when Firebase CLI is authenticated via GOOGLE_APPLICATION_CREDENTIALS (#6190)
 - Fixed bug where some extension instance updates would default to the wrong location.